LibWeb: Check if unloading is canceled without a traversable
Make checking if unloading is canceled a free function next to the
other navigation algorithms in LocalNavigable.cpp. The traversable was
only ever an optional input to the algorithm, so the navigate
algorithm's unload check and the top-level close no longer route through
LocalTraversableNavigable to reach it.
